pyproject.toml
[tool.poetry]
name = "online-education"
version = "0.1.0"
description = ""
authors = ["Victor Volkov <vektorn1212@gmail.com>"]
readme = "README.md"
packages = [{include = "online_education"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
pillow = "^10.1.0"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
django-filter = "^23.3"
djangorestframework-simplejwt = "^5.3.0"
coverage = "^7.3.2"
drf-yasg = "^1.21.7"
stripe = "^7.6.0"
celery = "^5.3.6"
redis = "^5.0.1"
django-celery-beat = "^2.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
